Now let's create the first page: When someone clicks on Get Started. It will start their loan application process. It will be 6 Page form. We will make it one by one don't rush in creating it right away after this prompt. Because there will be a lot of conditional journey.
First step is to ask the mobile number of customer.
It will be same responsive section with background & with the header similar to home page.
 There should be glassmorphed card in which first message should be from Aveka to share the contact number. And then:
select the country code, enter mobile number and get otp button.
In real service we will connect this GET OTP to OTP service but for now keep default OTP as 9999.
When person clicks on Get OTP, there should be a number text field that should appear to enter OTP.
And now we will show user "Save & Continue" button
All the buttons should follow the structure of Get started button