Perfect. After the user has provided the OTP and clicked on Save & Continue, it will take them to Step 1 of application form. Name of this step will be Admission KYC. Aveka will ask user if he/she has the Academic Offer Letter. It will have two buttons: "Yes" & "No" Flow will be different for both these options.
If user clicks "Yes", it will ask user to either upload the offer letter or click picture of offer letter.
As soon as a user uploads the offer letter (either PDF, word doc or image),  there should appear a preview section to show the document uploaded.
Meanwhile Gemini AI will extract the information from the offer letter. Gemini has to understand the language of the offer letter and pull all these points:
- Name of the Student
- Admission University Name
- Admission Course
- Admission Level (Graduation, Post graduation, PhD, Masters etc.)
- Admission Fees
- Course Start Date
- Offer Letter: Conditional or uncoditional

Once Gemini extracts the information, we need to show it in the section below in the form of a table.
We need provide user an edit option to be able to edit these details if they are not proper and then save it after manually changing the details.

Last section will be, need to ask user to provide consent that all the details mentioned above are correct and capture date and time of consent.
Since now we have the name of the user from the offer letter, we can start addressing the user in Aveka's message with users first message.

Also keep Aveka's message empathetic and user friendly