Under work experience, Upload resume or share linkedIn profile is optional. hence allow user to move to next section "Are you currently working" without him uploading or sharing the link.
Are you currently working is a mandatory step.
Infact Industry and experience is mandatory under work experience.

Last thing, as soon as these steps are complete, we should provide summary of the extracted information via Gemini AI (PAN card/national ID and Resume/LinkedIN)and manually filled information for user to review, edit and save in case of discrepency on the next page. But this next page should not be added in the progress bar as a separate step