Now the Next step in Offer Letter "No" journey will be a recommendation page in the app that will have: Country wise recommended courses and college. A button to refresh the recommendation A button to submit the select universities for application University card one for each recommendation which will have: - Name of the university - Summary of the university - Name of the course - Duration of the course - Fees of the course - Details of test required - A checkbox to select the university
For recommendation we will have a backend where we will upload the name of the university. Gemini will go through the list, search for the latest courses and it's details from the internet, it will store in the data store and backend as well. This list will also have a criteria mentioned against each course.
Gemini will evaluate the preference, past experience, academic KYC to recommend the courses to the user

We will use the login button to provide login to this backend as an admin panel.
If someone will login through @africred.org domain, password will be Globcred@2025 and he can see the admin pannel. But if someone will login through gmail.com id, they can only see the application form for loan, current saved application form, and if someone has completed the application then based on the academic offer letter "Yes" or "No" Last screen of the application page