i think, there is some issue with personal KYC. Aveka message for passport is not loading on the page and processing button is not working.
I think this is because firestore doesn't allow to store file in the system. I have created the storage in the same project with the link
gs://loan-app-jd6qx.firebasestorage.app and a folder in it with the name user_documents.
You can use this folder to upload the document and link it to the main database